fix: tests (tag order is now alphabetical)
All tags are serialized in alphabetical order.

Motivation: comparing XML files is easier when the order of tags and
attribs is predictable.  OTOH, DTD specifies a certain non-alphabetical
order, so we might rather stick to it in the future.

Exception: tags added via Contributor classes always go last.

The exception is not a design decision, it's just easier from the
technical POV.  Could be changed if desired.